Wire the palette color cascade the format has always documented (#15)
`format/palettes.md` and CONCEPTS §3 both promise a render-time
cascade — "the node uses the colors in its `style`" only when it has
no `color_schema`, and "editing a palette is a single-point change;
every node using it updates on the next render". Neither was true.
`resolve_theme_colors` had no production caller: node fill, frame and
text came exclusively from `node.style.*`, and `map.palettes` reached
the renderer only for border glyph cycling. Themed maps looked right
because the miMind converter had baked resolved colors into `style.*`,
so the gap was invisible until you edited a palette and nothing moved.

Option (a) from the issue: wire it, rather than re-document the
feature as dormant. The data model, the format docs and the whole
migration story already assume it, and CODE_CONVENTIONS §5 does not
allow documenting a gap instead of closing it.

The cascade now has one home, `model/theme.rs`, holding
`resolve_theme_colors` plus the readers every projection pass goes
through: `node_background_color`, `node_frame_color`,
`node_text_color`, `node_title_color`, and `edge_theme_stroke_color`.
Each answers "palette group, else the matching `style` field", and
each is a borrow, so the per-node per-frame path allocates nothing.

`connections_colored` threads into the edge cascade at #26's existing
helpers rather than beside them: `MindEdge::body_color` /
`label_color` / `portal_endpoint_color` / `portal_endpoint_text_color`
each take the palette tier as a `themed` argument — the same shape
`portal_endpoint_color` already used for its endpoint tier — and
`MindMap` supplies it. It sits below an explicit
`glyph_connection.color` (a color the author named is not a theme's to
overrule) and above `edge.color` (which is never absent, so a tier
under it could never win). The source node governs, not the target:
of the 248 `parent_child` edges in the testament map, 229 carry their
`from_id` node's frame color and 5 their `to_id` node's.

`starts_at_root: false` now means what `ColorSchema` says it means —
the schema root resolves to no group at all and keeps its own `style`,
and its children at `level: 1` take `groups[0]`.

`ColorSchema.level` becomes `usize`, the type CONCEPTS already claimed
and the type a slice index wants. It was `i32` indexed as
`schema.level as usize`, so an authored `-1` wrapped to a huge index
and clamped silently to the last group — a typo that rendered as a
plausible color. Serialized-field consequence: a map with a negative
level no longer loads, which is the intended trade (§10 licenses it,
and there are no such maps), and `maptool convert --legacy` floors a
negative legacy level to 0 with a message so the converter cannot emit
a file its own loader refuses.

Text needed one more thing than fill and frame, because
`style.text_color` never reached the renderer either: sections carry
their color only through `TextRun`s. A run naming its own color still
wins — an authored per-word color must survive a retheme — but a run
that leaves `color` empty now inherits the node default, and a section
with no runs at all gets the default as an explicit region instead of
falling through to cosmic-text's black. That also makes `title` real:
for section 0 it colors the first hard-newline-delimited line. Gaps
between authored runs are deliberately *not* filled;
`rich_text_spans_from_regions` renders only covered clusters, so
filling them would make text appear that the author's runs left out.

The reverse sync learns the same rule. A run-less section is not a
region-less section any more, so `custom::sync` projects the model
through `tree_builder::section_default_regions` — the very function
the forward path uses — and compares against that. Without it every
apply would read the synthesized default as divergence and write a
phantom `TextRun` into the model.

`maps/palette_cascade.mindmap.json` is the new fixture: every node's
`style` is a sentinel deliberately unequal to its palette group, so
every assertion against it fails if the cascade is unwired. It covers
`starts_at_root` both ways, a level past the last group, an unthemed
node, and edges whose source nodes differ on `connections_colored`.
The existing `theme_demo` could not do this — it has no palettes at
all.

Consequence worth naming: `maps/testament.mindmap.json` will render
differently. Its `style.*` values have drifted from its schemas — 218
of 252 themed nodes carry `#141414` where their palette group says a
pastel, and all 252 carry `#ffffff` text against a `#000000` group —
so the map now draws from its palettes as its schemas ask. Text is
unaffected (all 523 runs name `#ffffff` explicitly) and only 9 nodes
draw a frame at all, so the visible change is node fills and edge
strokes. The `#141414` is very likely conversion damage rather than
authorship: it is exactly the `cBackgroundColor="20,20,20,255,1"` on
`<Content>` in `maps/testament.mind`, whose nodes carry
`bForceColor="0"`, `bOverrideGroupColor="0"` and an unset
`cGroupFill` — miMind painted them from the group. If the dark look is
wanted back, it is now one edit per palette, which is the feature.

One documentation contradiction is resolved rather than silently
picked: palettes.md's "What `level` means" said a 7-group palette
themes 7 levels "before wrapping" while its resolution section, the
`ColorSchema` docs, CONCEPTS and the code all said clamp. The
miMind corpus agrees with the wrapping sentence — testament's `sandy`
palette has 7 groups and its nodes at levels 7-15 carry baked frames
matching `groups[level % 7]`. Clamp is kept, since it is what four of
the five sources and the implementation say; the sentence is corrected
and the evidence recorded in palettes.md so the choice is visible
rather than lost.

+++ lib/baumhard/src/mindmap/model/node.rs
@@ -722,33 +722,51 @@ pub struct NodeLayout {
 
 /// Links a node to one entry in a named [`super::Palette`] keyed by
 /// depth. `level` is the index into the palette's `groups`; clamped
-/// at theme-resolve time (`resolve_theme_colors`) so a schema
-/// referencing a level beyond the palette's length falls back to
-/// the last group rather than erroring.
+/// at theme-resolve time ([`super::MindMap::resolve_theme_colors`])
+/// so a schema referencing a level beyond the palette's length falls
+/// back to the last group rather than erroring.
 /// `starts_at_root` and `connections_colored` are round-tripped
-/// miMind-compat flags that the renderer interprets when resolving
-/// effective colors. Plain data; no runtime cost.
+/// miMind-compat flags that the projection passes interpret when
+/// resolving effective colors — see
+/// [`crate::mindmap::model::theme`] for the whole cascade. Plain
+/// data; no runtime cost.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ColorSchema {
     /// Named palette to bind this node's colors to — keys into
     /// [`super::MindMap::palettes`].
     pub palette: String,
-    /// Index into the palette's `groups` for this node's depth.
-    /// Clamped against the palette's length at resolve time so a
-    /// level past the end falls back to the last group.
-    pub level: i32,
+    /// Depth from the schema root, and the index into the palette's
+    /// `groups` it produces. Clamped against the palette's length at
+    /// resolve time so a level past the end falls back to the last
+    /// group.
+    ///
+    /// `usize` rather than a signed type because depth cannot be
+    /// negative and this value is only ever used as a slice index —
+    /// the `as usize` this field used to need turned an authored
+    /// `-1` into a huge index that then clamped silently to the last
+    /// group. A negative `level` is now refused by the loader with
+    /// serde's own message instead.
+    pub level: usize,
     /// `true` when depth indexing begins at the root (level 0 is the
     /// root itself); `false` shifts the indexing so the root is
-    /// transparent and children start at level 0.
+    /// transparent — it keeps its `style` colors — and its children,
+    /// at `level: 1`, take `groups[0]`.
     pub starts_at_root: bool,
-    /// When `true`, outgoing connections inherit the palette's
-    /// color at this node's level instead of the edge's own color.
+    /// When `true`, connections *leaving* this node inherit the
+    /// palette's `frame` color at this node's level instead of the
+    /// edge's own `color`. Read through
+    /// [`super::MindMap::edge_theme_stroke_color`].
     pub connections_colored: bool,
 }
 
 /// One palette entry — the four colors a themed node inherits at a
 /// given depth level. Referenced from [`ColorSchema::level`] via
-/// [`super::Palette::groups`]. Plain data; no runtime cost.
+/// [`super::Palette::groups`], resolved by
+/// [`super::MindMap::resolve_theme_colors`], and also the source of
+/// the per-glyph cycle a [`GlyphBorderConfig::color_palette`] border
+/// draws from (which channel it cycles is
+/// [`GlyphBorderConfig::color_palette_field`]). Plain data; no
+/// runtime cost.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ColorGroup {
     /// Background-fill color for a node at this level.
@@ -757,8 +775,15 @@ pub struct ColorGroup {
     pub frame: String,
     /// Text color for a node at this level.
     pub text: String,
-    /// First-line / title color — overrides `text` for the first
-    /// line of a node's text when present.
+    /// First-line / title color — stands in for `text` on the first
+    /// hard-newline-delimited line of the node's first
+    /// [`MindSection`], the stratum miMind called the node title.
+    /// Empty means "no distinct title color"; the whole section then
+    /// takes `text`.
+    ///
+    /// Like `text`, it is a *section default*: a [`TextRun`] naming
+    /// its own color keeps it, so a node whose runs all carry
+    /// explicit colors shows neither channel.
     pub title: String,
 }
